Updates to file transfer
Checks are now in place to see whether the camera is in automatic or manual capture mode. When in manual capture mode, FTP/dropbox data transfer is paused.

NOT CHECKED ON PI! This should be a fix for issue #24, and resulted in some changes associated with #9 updates - the check_acq_mode() function has been moved from check_run.py to a static method of CameraSpecs() so that it is more widely accessible.