Fix 140 and small edits to calibration series
Fixes #140 by catching preloaded calibration at the start of watching, raising an error to the screen and returning.
Also made some small edits to self.calibration_series usage which means it doesn't silently hit an error if preloaded is selected but no series is defined. Instead this sends a warning to the GUI, telling the user to select a calibration file.