iFit ILS fit option added
This was to some extent there already, but if an ILS_path was specified there was no way to revert to fitting the ILS as part of iFit. Now a `include_ils_fit` flag has been added, so that a user can switch between the two.

TODO - add GUI option to toggle ILS fit on and off
TODO - add config file option to load `include_ils_fit` flag