Subtract 0 if dark image can't be found
Subtracting a 0 from clear sky images if a dark image can't be found. This is just pulling BG images in line with what is done for plume images. It means that clear-sky images will be flagged as dark subtracted, so the clear sky can be used to correct the plume images (otherwise pyplis won't correct the images)