Update that makes ILS plot when fit option changed
iFit retrieval is now run when "Make ILS a fitted parameter" is selected and deselected in DOAS calibration panel. This means that the user can see the updated retrieval immediately.

Appears to be a bug whereby the initial retrieved column amount differs from the column amount you get if you select, then deselect this option. The difference is quite small, but there shouldn't be any difference when reverting back to the same processing approach. My guess is this relates to dark subtraction or how the ILS is initially read in.